split modvege_lib
